==> 01_borders.R <==
if (city=="DC"){
  
  at_large<-at_large_percent
  downtown<-downtown_percent+at_large
  greater_central<-greater_central_percent+downtown
  metro<-metro_percent+greater_central

  # if (metro != 1) {
  #   stop("Error: Sum of boroughs does not equal 100%")
  # }
  
  at_large_coords <- matrix(c(
    -77.119419, 38.934718,
    -77.040882, 38.995841,
    -76.909394, 38.892852,
    -77.029331, 38.799234, 
    -77.027054, 38.855403, 
    -77.059426, 38.899524, 
    -77.092116, 38.906666, 
    -77.119419, 38.934718
  ), ncol = 2, byrow = TRUE)

  downtown_coords <- matrix(c(
    -77.045444, 38.911450, 
    -77.003304, 38.896285, 
    -77.003261, 38.885795, 
    -77.033344, 38.884692, 
    -77.039652, 38.887867, 
    -77.047162, 38.884158, 
    -77.056432, 38.893812, 
    -77.056861, 38.902398, 
    -77.045444, 38.911450 
  ), ncol = 2, byrow = TRUE)

  greater_central_coords <- matrix(c(
    -77.071324, 38.904877,
    -77.071323, 38.942876,
    -77.027723, 38.945969,
    -77.000524, 38.931387,
    -76.983436, 38.899866,
    -76.988088, 38.877886,
    -77.020513, 38.875749,
    -77.050640, 38.886504,
    -77.057420, 38.901337,
    -77.071324, 38.904877
  ), ncol = 2, byrow = TRUE)
  
}

if (city=="NYC"){
  
  manhattan<-manhattan_percent
  brooklyn<-brooklyn_percent+manhattan
  queens<-queens_percent+brooklyn
  bronx<-bronx_percent+queens
  subway<-subway_percent+bronx
  
  if (subway != 1) {
    stop("Error: Sum of boroughs does not equal 100%")
  }
  
  manhattan_coords <- matrix(c(
    -74.013875, 40.700163, 
    -74.024023, 40.704360, 
    -74.014279, 40.755175, 
    -73.927928, 40.878489, 
    -73.910015, 40.872770, 
    -73.934220, 40.834756, 
    -73.933159, 40.808281, 
    -73.927902, 40.796740, 
    -73.942641, 40.782089, 
    -73.940642, 40.775750, 
    -73.970457, 40.742866, 
    -73.969927, 40.728005, 
    -73.975405, 40.710730, 
    -74.013875, 40.700163 
  ), ncol = 2, byrow = TRUE)

  brooklyn_coords <- matrix(c(
    -73.962692, 40.737428, 
    -73.929043, 40.727534, 
    -73.921881, 40.709394, 
    -73.895402, 40.681704, 
    -73.868821, 40.694847, 
    -73.855515, 40.642089, 
    -73.911528, 40.577762, 
    -74.013005, 40.570018, 
    -74.013976, 40.583109, 
    -74.001595, 40.582741, 
    -74.006693, 40.596383, 
    -74.031212, 40.603387, 
    -74.045535, 40.620895, 
    -74.018345, 40.663263, 
    -74.021016, 40.679282, 
    -73.996496, 40.704684, 
    -73.971977, 40.709469, 
    -73.962692, 40.737428 
  ), ncol = 2, byrow = TRUE)

  queens_coords <- matrix(c(
    -73.967564, 40.736458,
    -73.931842, 40.779854,
    -73.909949, 40.792220,
    -73.888406, 40.781261,
    -73.887799, 40.767105,
    -73.873233, 40.771977,
    -73.850899, 40.759014,
    -73.855754, 40.781445,
    -73.861581, 40.786224,
    -73.837061, 40.801662,
    -73.771879, 40.798630,
    -73.755856, 40.768024,
    -73.751608, 40.783926,
    -73.701719, 40.752578,
    -73.700263, 40.739427,
    -73.707789, 40.728022,
    -73.730123, 40.722135,
    -73.725025, 40.652367,
    -73.743839, 40.643618,
    -73.821883, 40.666261,
    -73.822597, 40.647909,
    -73.853239, 40.643285,
    -73.869289, 40.694845,
    -73.896493, 40.682682,
    -73.922375, 40.709145,
    -73.928582, 40.728072,
    -73.967564, 40.736458
  ), ncol = 2, byrow = TRUE)

  bronx_coords <- matrix(c(
    -73.910848, 40.915267,
    -73.819450, 40.888649,
    -73.790025, 40.804987,
    -73.912296, 40.796159,
    -73.928312, 40.802131,
    -73.933776, 40.810322,
    -73.934095, 40.834624,
    -73.909793, 40.872164,
    -73.925643, 40.878535,
    -73.910848, 40.915267
  ), ncol = 2, byrow = TRUE)
  
  if (subway_percent>0){
    temp_file <- tempfile(fileext = ".zip")
    download.file("http://web.mta.info/developers/data/nyct/subway/google_transit.zip", temp_file)
  
    gtfs_data <- read_gtfs(temp_file)
  }

}
